// src/lib/auth/auth.graphql.d.ts
export declare type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
};
export declare type Auth = {
    __typename?: 'Auth';
    isAuthenticated: Scalars['Boolean'];
    user: User;
    userFullName: Scalars['String'];
};
export declare type User = {
    __typename?: 'User';
    gender: Scalars['String'];
    name: UserName;
};
export declare type UserName = {
    __typename?: 'UserName';
    title: Scalars['String'];
    first: Scalars['String'];
    last: Scalars['String'];
};
export declare type Query = {
    __typename?: 'Query';
    viewer: OldUser;
    auth: Auth;
};
export declare type Mutation = {
    __typename?: 'Mutation';
    updateName: OldUser;
    signIn: Auth;
};
export declare type MutationUpdateNameArgs = {
    name: Scalars['String'];
};
export declare type MutationSignInArgs = {
    name?: Maybe<Scalars['String']>;
};
export declare type OldUser = {
    __typename?: 'OldUser';
    id: Scalars['ID'];
    name: Scalars['String'];
    status: Scalars['String'];
};
export declare type AuthQueryVariables = Exact<{
    [key: string]: never;
}>;
export declare type AuthQuery = ({
    __typename?: 'Query';
} & {
    auth: ({
        __typename?: 'Auth';
    } & Pick<Auth, 'isAuthenticated' | 'userFullName'> & {
        user: ({
            __typename?: 'User';
        } & Pick<User, 'gender'> & {
            name: ({
                __typename?: 'UserName';
            } & Pick<UserName, 'first'>);
        });
    });
});
export declare type SignInMutationVariables = Exact<{
    name?: Maybe<Scalars['String']>;
}>;
export declare type SignInMutation = ({
    __typename?: 'Mutation';
} & {
    signIn: ({
        __typename?: 'Auth';
    } & Pick<Auth, 'isAuthenticated' | 'userFullName'>);
});
export declare const AuthDocument: Apollo.DocumentNode;